introduce a separate RenderAttributeValue trait, fixes #28
This is a breaking change if you were rendering templates as attribute
values. #28 shows how this could easily generate invalid markup without
the use of `markup::raw`, which is something I want to prevent.

If you want to render a template as attribute value now, you'll have to
wrap it in `markup::raw()`:

```rust
markup::define! {
    A { #a {} }
    B { div[a = markup::raw(A {})] {} }
}

fn main() {
    println!("{}", B {});
}
```

prints:

```
<div a="<div id="a"></div>"></div>
```